refactor: isolate agent root namespace mechanics
